Re-upgrade jekyll etc. (PLATFORM-4085)
This reverts commit 7ceeab3f2a8616604f7825d38bea82d3e04fe78d.
