+++ Gemfile
@@ -4,14 +4,14 @@ gem 'rake'
 gem 'compass'
 gem 'jekyll'
 gem 'octopress-hooks'
-gem 'rdiscount'
-gem 'pygments.rb'
-gem 'RedCloth'
+gem 'kramdown-parser-gfm'
 gem 'pry'
 gem 'mp3info'
-gem 'aws-sdk'
+gem 'aws-sdk-s3'
 gem 'jekyll-lunr-js-search', git: 'https://github.com/ashfurrow/jekyll-lunr-js-search.git', branch: 'v2'
 gem 'jekyll-redirect-from'
+gem 'jekyll-paginate'
+gem 'byebug'
 
 group :distribution do
   gem 'danger'
