+++ _config.yml
@@ -15,15 +15,15 @@ disqus_show_comment_count: false
 
 exclude: ["vendor", "LICENSE", "_posts/LICENSE"]
 
-gems: ["jekyll-lunr-js-search", "jekyll-redirect-from"]
+plugins: ["jekyll-lunr-js-search", "jekyll-redirect-from", "jekyll-paginate"]
 lunr_search:
   js_dir: "javascripts"
 
-markdown: rdiscount
-highlighter: pygments
-rdiscount:
-  extensions:
-    - generate_toc
+markdown: kramdown
+highlighter: rouge
+kramdown:
+  input: GFM
+  hard_wrap: false
 
 paginate: 10
 paginate_path: "/blog/:num/"
