+++ _includes/article.html
@@ -1,8 +1,4 @@
-{% if page.single == true %}
 <div class="article-container-single {{ page.article-class }}">
-{% else %}
-<div class="article-container">
-{% endif %}
 
   {% unless page.no_header %}
     <div class="meta-container">
@@ -35,43 +31,41 @@
       <div class="entry-content">{{ content }}</div>
     {% endif %}
 
-    {% if page.single == true %}
-      <p class="meta-author">
-        {% include post/author.html %}
-      </p>
-      <p class="meta-categories">
-        Categories: {% include post/categories.html %}
-      </p>
-      {% if page.series %}
-       <p class="meta-series" style="padding-bottom:0;">Part of a series: <strong>{{ page.series | series_link }}</strong></p>
-       <ul>
-         {% for post in site.posts %}
-          {% if post.series == page.series %}
-          <li>{% include archive_post.html %}</li>
-          {% endif %}
-        {% endfor %}
-
-       </ul>
-      {% endif %}
-
-      {% if site.disqus_short_name and page.comments == true %}
-        <hr/>
-        <section class="comments">
-          <h1>Comments</h1>
-          <div id="disqus_thread" aria-live="polite">{% include post/disqus_thread.html %}</div>
-        </section>
-        <hr/>
-      {% endif %}
-
-      <p class="meta-paginate">
-        {% if page.previous.url %}
-          <a class="previous" href="{{page.previous.url}}" title="Previous Post: {{page.previous.title}}">&laquo; {{page.previous.title}}</a>
+    <p class="meta-author">
+      {% include post/author.html %}
+    </p>
+    <p class="meta-categories">
+      Categories: {% include post/categories.html %}
+    </p>
+    {% if page.series %}
+      <p class="meta-series" style="padding-bottom:0;">Part of a series: <strong>{{ page.series | series_link }}</strong></p>
+      <ul>
+        {% for post in site.posts %}
+        {% if post.series == page.series %}
+        <li>{% include archive_post.html %}</li>
         {% endif %}
-        {% if page.next.url %}
-          <a class="next" href="{{page.next.url}}" title="next Post: {{page.next.title}}">{{page.next.title}} &raquo;</a>
-        {% endif %}
-      </p>
+      {% endfor %}
+
+      </ul>
     {% endif %}
+
+    {% if site.disqus_short_name and page.comments == true %}
+      <hr/>
+      <section class="comments">
+        <h1>Comments</h1>
+        <div id="disqus_thread" aria-live="polite">{% include post/disqus_thread.html %}</div>
+      </section>
+      <hr/>
+    {% endif %}
+
+    <p class="meta-paginate">
+      {% if page.previous.url %}
+        <a class="previous" href="{{page.previous.url}}" title="Previous Post: {{page.previous.title}}">&laquo; {{page.previous.title}}</a>
+      {% endif %}
+      {% if page.next.url %}
+        <a class="next" href="{{page.next.url}}" title="next Post: {{page.next.title}}">{{page.next.title}} &raquo;</a>
+      {% endif %}
+    </p>
   </div>
