+++ _includes/article_list_item.html
@@ -1,0 +1,39 @@
+<div class="article-container">
+
+  {% unless page.no_header %}
+    <div class="meta-container">
+      <header>
+        {% unless page.meta == false %}
+          <a href="{{ root_url }}{{ post.url }}">
+        {% endunless %}
+        {% if index %}
+          <h2 class="entry-title">{% if site.titlecase %}{{ post.title | titlecase }}{% else %}{{ post.title }}{% endif %}</h2>
+        {% else %}
+          <h2 class="entry-title">{% if site.titlecase %}{{ page.title | titlecase }}{% else %}{{ page.title }}{% endif %}</h2>
+        {% endif %}
+        <time datetime="{{ page.date | date: "%Y-%m-%d" }}">{{ page.date | date: "%b %-d, %Y" }}</time>
+          <p class="meta-author">{% include post/mini-author.html %}</p>
+          </a>
+      </header>
+    </div>
+  {% endunless %}
+
+  <div class="content-container">
+    {% if index %}
+    {% if post.content contains '<!-- more -->' %}
+        {{ post.content | split:'<!-- more -->' | first }}
+        <p><a href="{{ post.url }}">Read on →</a></p>
+    {% else %}
+        {{ post.content }}
+    {% endif %}
+
+    {% else %}
+      <div class="entry-content">{{ content }}</div>
+    {% endif %}
+  </div>
+
+
+  <footer>
+  </footer>
+
+</div>
