+++ _includes/epic_img.html
@@ -1,5 +1,3 @@
-</article>
-<a href='{{ include.url }}' style="{{ include.astyle }}">
+<a href="{{ include.url }}" style="{{ include.astyle }}">
   <img src="{{ include.url }}" alt="{{ include.title }}" style="{{ include.style }}">
 </a>
-<article class="post">
