+++ _includes/expanded_img.html
@@ -1,5 +1,3 @@
-</div></div>
-<a href='{{ include.url }}'>
+<a href="{{ include.url }}">
   <img src="{{ include.url }}" alt="{{ include.title }}" style="{{ include.style }}">
 </a>
-<div class='meta-container'><header>&nbsp;</header></div><div class='content-container'><div class='entry-content'>
