+++ _layouts/category_index.html
@@ -4,7 +4,7 @@ footer: false
 ---
 
 <div id="blog-archives" class="category">
-{% for post in site.categories[page.category] %}
+{% for post in site.posts %} {% if post.categories contains page.category.name %}
 {% capture this_year %}{{ post.date | date: "%Y" }}{% endcapture %}
 {% unless year == this_year %}
   {% assign year = this_year %}
@@ -13,5 +13,5 @@ footer: false
 <article>
   {% include archive_post.html %}
 </article>
-{% endfor %}
+{% endif %} {% endfor %}
 </div>
