+++ _layouts/epic.html
@@ -60,7 +60,7 @@
 
           <section>
             <h4>Author</h4>
-            <p>{% for key in page.author %}{% assign author = site.authors[key] %}{% if forloop.index > 1 %},{% endif %} <a href="/author/{{key}}"/>{{ author.name }}</a>{% endfor %}</p>
+            <p>{% for key in page.author %}{% assign author = site.authors[key] %}{% if forloop.index > 1 %},{% endif %} <a href="/author/{{key}}">{{ author.name }}</a>{% endfor %}</p>
             <p>{% for key in page.author %}{% assign author = site.authors[key] %}{{ author.byline }}</a>{% endfor %}</p>
           </section>
 
@@ -86,7 +86,7 @@
               <li><a href='http://artsy.github.io/blog/archives/'>Archives</a></li>
               <li><a href='http://artsy.github.io/search/'>Search</a></li>
               <li><a href='https://github.com/artsy/artsy.github.io'>Code on GitHub</a></li>
-              <li><a href='https://github.com/artsy/artsy.github.io/edit/source/{{page.path}}'>Fix typoes in this post</a></li>
+              <li><a href='https://github.com/artsy/artsy.github.io/edit/source/{{page.path}}'>Fix typos in this post</a></li>
             </ul>
           </section>
