+++ _posts/2015-12-11-License-and-You.markdown
@@ -16,7 +16,7 @@ Jump [to YouTube](https://www.youtube.com/watch?v=0r7JcN3Q_LY) for the video, or
 
 <!-- more -->
 
-{% youtube 0r7JcN3Q_LY %}
+<iframe width="100%" height="400" src="https://www.youtube.com/embed/0r7JcN3Q_LY" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
 
 ### Licenses For Code
