+++ _posts/2016-02-11-Code-Review-Energy.markdown
@@ -15,4 +15,4 @@ Jump [to YouTube](https://www.youtube.com/watch?v=Xhd25hFzN4o) for the video, or
 
 <!-- more -->
 
-{% youtube Xhd25hFzN4o %}
+<iframe width="100%" height="400" src="https://www.youtube.com/embed/Xhd25hFzN4o" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
