+++ _posts/2016-03-05-iOS-Code-Injection.markdown
@@ -16,4 +16,4 @@ Jump [to YouTube](https://www.youtube.com/watch?v=uftvtmyZ8TM) for the video, or
 
 <!-- more -->
 
-{% youtube uftvtmyZ8TM %}
+<iframe width="100%" height="400" src="https://www.youtube.com/embed/uftvtmyZ8TM" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
