+++ _posts/2016-08-23-swift-at-artsy-beginners-lesson-two.markdown
@@ -13,8 +13,7 @@ Yesterday was lesson two and I have a [screen recording][youtube] of my walkthro
 
 <!-- more -->
 
-{% youtube EBtzqx1YRBg %}
-
+<iframe width="100%" height="400" src="https://www.youtube.com/embed/EBtzqx1YRBg" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
 
 [course]: http://artsy.github.io/blog/2016/01/26/swift-at-artsy/
 [youtube]: https://youtu.be/EBtzqx1YRBg
