+++ _posts/2016-09-16-Dev-Interview-Maxim-Cramer.markdown
@@ -15,4 +15,4 @@ Jump [to YouTube](https://www.youtube.com/watch?v=ftsBR_KqHcE) for the video, or
 
 <!-- more -->
 
-{% youtube ftsBR_KqHcE %}
+<iframe width="100%" height="400" src="https://www.youtube.com/embed/ftsBR_KqHcE" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
