+++ _posts/2018-03-18-ar.markdown
@@ -23,13 +23,6 @@ license.
 
 <!-- more -->
 
-<div class="mobile-only">
-<p>
-  <strong>Before you get started</strong>, it looks like you're using a really small screen, 
-  this post is built with larger screens in mind. You will be missing interactable elements which you <strong>really should see</strong>.
-</p>
-</div>
-
 ## How does ARKit work?
 
 You start up ARKit by creating an [AR Session][ar_session] and passing it to an view on your screen. This will trigger
@@ -44,7 +37,7 @@ at this as a diagram:
     <li>ARSession starts</li>
     <li>It detects some feature points</li>
     <li>You can then attach objects to those points</li>
-  </ul>
+  </ol>
 </div>
 <div style='flex:1'>{% include svg/ar/what-is-ar.svg %}</div>
 </article>
@@ -89,7 +82,7 @@ Here's how our original approach worked:
     <li>Detect the edge of a floor by asking a user to point at the floor near the wall</li>
     <li>Find the edge furthest away</li>
     <li>Attach a work to the floor closest to the wall</li>
-  </ul>
+  </ol>
 </div>
 <div style='flex:1'>{% include svg/ar/v1.svg %}</div>
 </article>
@@ -106,7 +99,7 @@ From there, we experimented with using the location of the iPhone as a point of 
   <li>Ask a user to put their phone against the wall</li>
   <li>When the camera is fully osbscured, use the last known point as an anchor</li>
   <li>When the user pulls back and reconnects, place the work</li>
-  </ul>
+  </ol>
 </div>
 <div style='flex:1'>{% include svg/ar/v2.svg %}</div>
 </article>
@@ -132,7 +125,7 @@ should be:
     <li>Ask a user to point at a wall</li>
     <li>Use the feature detection to detect a plane</li>
     <li>Use that plane to place an artwork</li>
-  </ul>
+  </ol>
 </div>
 <div style='flex:1'>{% include svg/ar/v3.svg %}</div>
 </article>
@@ -148,7 +141,7 @@ to end up in frustration. So instead, our user flow is:
     <li>Ask a user to point at an existing object on a wall</li>
     <li>Expand the discovered plane in all directions</li>
     <li>Use the extended plane to place an artwork</li>
-  </ul>
+  </ol>
 </div>
 <div style='flex:1'>{% include svg/ar/v4.svg %}</div>
 </article>
