+++ blog/archives/index.html
@@ -10,7 +10,7 @@ exclude_from_search: true
   {% capture this_year %}{{ post.date | date: "%Y" }}{% endcapture %}
   {% unless year == this_year %}
     {% assign year = this_year %}
-    <h2 id="{{year}}"">{{ year }}</h2>
+    <h2 id="{{year}}">{{ year }}</h2>
   {% endunless %}
 
   <article>
@@ -21,7 +21,7 @@ exclude_from_search: true
         {% for key in author_key %}
           {% assign author = site.authors[key] %}
           {% if author %}
-              <span class="byline author vcard"><a href ="/author/{{key}}">{{ author.name }}<a/></span>
+              <span class="byline author vcard"><a href ="/author/{{key}}">{{ author.name }}</a></span>
           {% endif %}
         {% endfor %}
       </a>
