+++ index.html
@@ -7,7 +7,7 @@ layout: default
   {% for post in paginator.posts %}
     {% assign content = post.content %}
       <article>
-        {% include article.html %}
+        {% include article_list_item.html %}
       </article>
   {% endfor %}
